moved devices and restored dashboard
